+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
  - Custom Proton installations now use Steam Runtime installations when applicable
  - Fix crash caused by older Steam app installations using a different app manifest structure
- - Fix crash caused by the change to lowercase field names in multiple VDF files
+ - Fix crash caused by change to lowercase field names in multiple VDF files
+ - Fix crash caused by change in the Steam library folder configuration file
 
 ## [1.5.1] - 2021-05-10
 ### Fixed